+++ extensions/mdn-docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MDN Docs Changelog
 
-## [Search Through MDN Docs] - 2022-01-17
+## [Search Through MDN Docs] - 2023-02-11
 
--   You can search through the MDN docs
+-   [feat] You can search through the MDN docs
+
+## [History & Clipboard] - 2023-02-12
+
+-   [feat] Copy url to clipboard on a search result
+-   [feat] Recently searched results are saved to history, can clear all recently searched items at once or by item